+++ components/microsoft_dynamics_365_sales/common/language-codes.mjs
@@ -1,0 +1,434 @@
+export default [
+  {
+    label: "Afrikaans - South Africa",
+    value: 1078,
+  },
+  {
+    label: "Albanian - Albania",
+    value: 1052,
+  },
+  {
+    label: "Arabic - Algeria",
+    value: 5121,
+  },
+  {
+    label: "Arabic - Bahrain",
+    value: 15361,
+  },
+  {
+    label: "Arabic - Egypt",
+    value: 3073,
+  },
+  {
+    label: "Arabic - Iraq",
+    value: 2049,
+  },
+  {
+    label: "Arabic - Jordan",
+    value: 11265,
+  },
+  {
+    label: "Arabic - Kuwait",
+    value: 13313,
+  },
+  {
+    label: "Arabic - Lebanon",
+    value: 12289,
+  },
+  {
+    label: "Arabic - Libya",
+    value: 4097,
+  },
+  {
+    label: "Arabic - Morocco",
+    value: 6145,
+  },
+  {
+    label: "Arabic - Oman",
+    value: 8193,
+  },
+  {
+    label: "Arabic - Qatar",
+    value: 16385,
+  },
+  {
+    label: "Arabic - Saudi Arabia",
+    value: 1025,
+  },
+  {
+    label: "Arabic - Syria",
+    value: 10241,
+  },
+  {
+    label: "Arabic - Tunisia",
+    value: 7169,
+  },
+  {
+    label: "Arabic - U.A.E.",
+    value: 14337,
+  },
+  {
+    label: "Arabic - Yemen",
+    value: 9217,
+  },
+  {
+    label: "Armenian - Armenia",
+    value: 1067,
+  },
+  {
+    label: "Azeri (Cyrillic) - Azerbaijan",
+    value: 2092,
+  },
+  {
+    label: "Azeri (Latin) - Azerbaijan",
+    value: 1068,
+  },
+  {
+    label: "Basque - Spain",
+    value: 1069,
+  },
+  {
+    label: "Belarusian - Belarus",
+    value: 1059,
+  },
+  {
+    label: "Bulgarian - Bulgaria",
+    value: 1026,
+  },
+  {
+    label: "Catalan - Spain",
+    value: 1027,
+  },
+  {
+    label: "Chinese - Hong Kong S.A.R.",
+    value: 3076,
+  },
+  {
+    label: "Chinese - Macau S.A.R.",
+    value: 5124,
+  },
+  {
+    label: "Chinese - People's Republic of China",
+    value: 2052,
+  },
+  {
+    label: "Chinese - Singapore",
+    value: 4100,
+  },
+  {
+    label: "Chinese - Taiwan",
+    value: 1028,
+  },
+  {
+    label: "Croatian - Croatia",
+    value: 1050,
+  },
+  {
+    label: "Czech - Czech Republic",
+    value: 1029,
+  },
+  {
+    label: "Danish - Denmark",
+    value: 1030,
+  },
+  {
+    label: "Divehi - Maldives",
+    value: 1125,
+  },
+  {
+    label: "Dutch - Belgium",
+    value: 2067,
+  },
+  {
+    label: "Dutch - Netherlands",
+    value: 1043,
+  },
+  {
+    label: "English - Australia",
+    value: 3081,
+  },
+  {
+    label: "English - Belize",
+    value: 10249,
+  },
+  {
+    label: "English - Canada",
+    value: 4105,
+  },
+  {
+    label: "English - Caribbean",
+    value: 9225,
+  },
+  {
+    label: "English - Ireland",
+    value: 6153,
+  },
+  {
+    label: "English - Jamaica",
+    value: 8201,
+  },
+  {
+    label: "English - New Zealand",
+    value: 5129,
+  },
+  {
+    label: "English - Republic of the Philippines",
+    value: 13321,
+  },
+  {
+    label: "English - South Africa",
+    value: 7177,
+  },
+  {
+    label: "English - Trinidad and Tobago",
+    value: 11273,
+  },
+  {
+    label: "English - United Kingdom",
+    value: 2057,
+  },
+  {
+    label: "English - United States",
+    value: 1033,
+  },
+  {
+    label: "English - Zimbabwe",
+    value: 12297,
+  },
+  {
+    label: "Estonian - Estonia",
+    value: 1061,
+  },
+  {
+    label: "Faroese - Faeroe Islands",
+    value: 1080,
+  },
+  {
+    label: "Farsi - Iran",
+    value: 1065,
+  },
+  {
+    label: "Finnish - Finland",
+    value: 1035,
+  },
+  {
+    label: "French - Belgium",
+    value: 2060,
+  },
+  {
+    label: "French - Canada",
+    value: 3084,
+  },
+  {
+    label: "French - France",
+    value: 1036,
+  },
+  {
+    label: "French - Luxembourg",
+    value: 5132,
+  },
+  {
+    label: "French - Monaco",
+    value: 6156,
+  },
+  {
+    label: "French - Switzerland",
+    value: 4108,
+  },
+  {
+    label: "Macedonian - North Macedonia",
+    value: 1071,
+  },
+  {
+    label: "Galician - Spain",
+    value: 1110,
+  },
+  {
+    label: "Georgian - Georgia",
+    value: 1079,
+  },
+  {
+    label: "German - Austria",
+    value: 3079,
+  },
+  {
+    label: "German - Germany",
+    value: 1031,
+  },
+  {
+    label: "German - Liechtenstein",
+    value: 5127,
+  },
+  {
+    label: "German - Luxembourg",
+    value: 4103,
+  },
+  {
+    label: "German - Switzerland",
+    value: 2055,
+  },
+  {
+    label: "Greek - Greece",
+    value: 1032,
+  },
+  {
+    label: "Gujarati - India",
+    value: 1095,
+  },
+  {
+    label: "Hebrew - Israel",
+    value: 1037,
+  },
+  {
+    label: "Hindi - India",
+    value: 1081,
+  },
+  {
+    label: "Hungarian - Hungary",
+    value: 1038,
+  },
+  {
+    label: "Icelandic - Iceland",
+    value: 1039,
+  },
+  {
+    label: "Indonesian - Indonesia",
+    value: 1057,
+  },
+  {
+    label: "Italian - Italy",
+    value: 1040,
+  },
+  {
+    label: "Italian - Switzerland",
+    value: 2064,
+  },
+  {
+    label: "Japanese - Japan",
+    value: 1041,
+  },
+  {
+    label: "Kannada - India",
+    value: 1099,
+  },
+  {
+    label: "Kazakh - Kazakhstan",
+    value: 1087,
+  },
+  {
+    label: "Korean - Korea",
+    value: 1042,
+  },
+  {
+    label: "Latvian - Latvia",
+    value: 1062,
+  },
+  {
+    label: "Lithuanian - Lithuania",
+    value: 1063,
+  },
+  {
+    label: "Malay - Brunei Darussalam",
+    value: 2110,
+  },
+  {
+    label: "Malay - Malaysia",
+    value: 1086,
+  },
+  {
+    label: "Marathi - India",
+    value: 1102,
+  },
+  {
+    label: "Mongolian - Mongolia",
+    value: 1104,
+  },
+  {
+    label: "Norwegian (Bokmål) - Norway",
+    value: 1044,
+  },
+  {
+    label: "Norwegian (Nynorsk) - Norway",
+    value: 2068,
+  },
+  {
+    label: "Polish - Poland",
+    value: 1045,
+  },
+  {
+    label: "Portuguese - Brazil",
+    value: 1046,
+  },
+  {
+    label: "Portuguese - Portugal",
+    value: 2070,
+  },
+  {
+    label: "Punjabi - India",
+    value: 1094,
+  },
+  {
+    label: "Romanian - Romania",
+    value: 1048,
+  },
+  {
+    label: "Russian - Russia",
+    value: 1049,
+  },
+  {
+    label: "Spanish - Spain",
+    value: 1034,
+  },
+  {
+    label: "Spanish - Mexico",
+    value: 2058,
+  },
+  {
+    label: "Spanish - Argentina",
+    value: 11274,
+  },
+  {
+    label: "Spanish - Chile",
+    value: 13322,
+  },
+  {
+    label: "Spanish - Colombia",
+    value: 9226,
+  },
+  {
+    label: "Swahili - Kenya",
+    value: 1089,
+  },
+  {
+    label: "Swedish - Sweden",
+    value: 1053,
+  },
+  {
+    label: "Tamil - India",
+    value: 1097,
+  },
+  {
+    label: "Thai - Thailand",
+    value: 1054,
+  },
+  {
+    label: "Turkish - Turkey",
+    value: 1055,
+  },
+  {
+    label: "Ukrainian - Ukraine",
+    value: 1058,
+  },
+  {
+    label: "Urdu - Pakistan",
+    value: 1056,
+  },
+  {
+    label: "Vietnamese - Vietnam",
+    value: 1066,
+  },
+  {
+    label: "Welsh - United Kingdom",
+    value: 1106,
+  },
+];