+++ components/opensrs/sources/new-dns-zone-change/test-event.mjs
@@ -1,0 +1,11 @@
+export default {
+  object: "DOMAIN",
+  event_id: "bf7b2cf6d8e835324241a4eae7be4ee5",
+  event: "ZONE_CHECK_STATUS_CHANGE",
+  event_date: "2014-05-27T19:29:25Z",
+  objectData: {
+    domain_name: "abc-fictional.de",
+    domain_id: "801743",
+    zone_check_status: "invalid",
+  },
+};