+++ packages/sdk/src/server/index.ts
@@ -4,7 +4,7 @@
 
 import * as oauth from "oauth4webapi";
 import {
-  Account, BaseClient, type AppInfo, type ConnectTokenResponse,
+  Account, BaseClient, type AppInfo, type ConnectTokenResponse, type RequestOptions,
 } from "../shared/index.js";
 export * from "../shared/index.js";
 
@@ -108,6 +108,48 @@ export type GetAccountByIdOpts = {
 };
 
 /**
+ * Options used to determine the external user and account to be used in Connect Proxy API
+ */
+export type ProxyApiOpts  = {
+  /**
+   * Search parameters to be added to the proxy request.  external_user_id and account_id are required.
+   */
+  searchParams: Record<string, string>;
+};
+
+/**
+ * fetch-like options for the Target of the Connect Proxy Api Request
+ */
+export type ProxyTargetApiOpts = {
+  /**
+   * http method for the request
+   */
+  method: "GET" | "POST" | "PUT" | "DELETE" | "PATCH";
+  /**
+   * http headers for the request
+   */
+  headers?: Record<string, string>;
+  /**
+   * http body for the request
+   */
+  body?: string;
+};
+
+/**
+ * object that contains the url and options for the target of the Connect Proxy Api Request
+ */
+export type ProxyTargetApiRequest = {
+  /**
+   * URL for the target of the request.  Search parameters must be included here.
+   */
+  url: string;
+  /**
+   * fetch-like options for the target of the Connect Proxy Request
+   */
+  options: ProxyTargetApiOpts;
+};
+
+/**
  * Creates a new instance of BackendClient with the provided options.
  *
  * @example
@@ -143,7 +185,7 @@ export class BackendClient extends BaseClient {
     token: string
     expiresAt: number
   };
-  protected override projectId: string;
+  protected override projectId: string = "";
 
   /**
    * Constructs a new ServerClient instance.
@@ -199,7 +241,11 @@ export class BackendClient extends BaseClient {
   }
 
   private async ensureValidOauthAccessToken(): Promise<string> {
-    const { client, clientAuth, as } = this.oauthClient
+    const {
+      client,
+      clientAuth,
+      as,
+    } = this.oauthClient
 
     let attempts = 0;
     const maxAttempts = 2;
@@ -221,7 +267,9 @@ export class BackendClient extends BaseClient {
           token: oauthTokenResponse.access_token,
           expiresAt: Date.now() + (oauthTokenResponse.expires_in || 0) * 1000,
         };
-      } catch {}
+      } catch {
+        // pass
+      }
 
       attempts++;
     }
@@ -367,5 +415,35 @@ export class BackendClient extends BaseClient {
     return this.makeConnectRequest("/projects/info", {
       method: "GET",
     });
+  }
+
+  /**
+   * Makes a proxy request to the target app API with the specified query parameters and options.
+   *
+   * @returns A promise resolving to the response from the downstream service
+   */
+  public makeProxyRequest(proxyOptions: ProxyApiOpts, targetRequest: ProxyTargetApiRequest): Promise<string> {
+    const url64 = btoa(targetRequest.url).replace(/\+/g, "-")
+      .replace(/\//g, "_")
+      .replace(/=+$/, "");
+
+    const headers = targetRequest.options.headers || {};
+
+    const newHeaders = Object.keys(headers).reduce<{ [key: string]: string }>((acc, key) => {
+      acc[`x-pd-proxy-${key}`] = headers[key];
+      return acc;
+    }, {});
+
+    const newOpts: RequestOptions = {
+      method: targetRequest.options.method,
+      headers: newHeaders,
+      params: proxyOptions.searchParams,
+    }
+
+    if (targetRequest.options.body) {
+      newOpts.body = targetRequest.options.body
+    }
+
+    return this.makeConnectRequest(`/proxy/${url64}`, newOpts);
   }
 }