+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 # Changelog
 
+## [1.3.3] - 2025-02-5
+
+### Changed
+
+- Add makeProxyRequest function to BaseClient
+
+## [1.3.2] - 2025-02-3
+
+### Changed
+
+- Add getEnvironment function to BaseClient
+
 ## [1.3.1] - 2025-01-30
 
 ### Changed